20190625:1、EQ更改；2、修复华为手机不能自动播放的问题；3、修复自动关机不成功的问题。

+++ apps/applications/earbud/av_headset_sm.h
@@ -200,7 +200,7 @@ typedef struct
 	
 #endif
 #ifdef	AUTO_POWER_OFF
-	uint8 auto_power_off_times;
+	uint16 auto_power_off_times;
 #endif
 } smTaskData;
